pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "aniemore"
version = "1.0.1"
authors = [
    "Ilya Lubenets <lii291001@gmail.com>",
    "Nikita Davidchuk <ar4ikov228@gmail.com>",
    "Artem Amentes <artem@socialcode.ru>",
]
description = "Aniemore (Artem Nikita Ilya EMOtion REcognition) is a library for emotion recognition in voice and text for russian language."
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Natural Language :: Russian",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Text Processing :: Linguistic",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
]

[tool.poetry.urls]
"Homepage" = "https://github.com/aniemore/Aniemore"
"Bug Tracker" = "https://github.com/aniemore/Aniemore/issues"
"Hugging Face Homepage" = "https://huggingface.co/Aniemore"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
torch = "^1.13.1"
transformers = "^4.26.1"
torchaudio = "^0.13.1"
numpy = "^1.24.2"
openai-whisper = "^20230124"
datasets = "^2.9.0"
coverage = {version = "^7.1.0", optional = true}

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"

[tool.pytest.ini_options]
pythonpath = [
".", "aniemore",
]
